+++ docs/scripts/initiators/version.initiator.js
@@ -1,0 +1,23 @@
+/**
+ * Generated header by Cozen for 21-days project
+ * Generated file version.initiator on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 02/01/2018
+ * Time: 11:45
+ * Version: 1.0.0
+ */
+(function (angular) {
+	'use strict';
+
+	angular
+		.module('21days')
+		.run(config);
+
+	config.$inject = [];
+
+	function config() {
+		console.info('21-days version: /* @echo version */');
+	}
+
+}(window.angular));