+++ grunt/config/preprocess.js
@@ -1,0 +1,22 @@
+/**
+ * Generated header by Cozen for 21-days project
+ * Generated file preprocess on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 02/01/2018
+ * Time: 11:52
+ * Version: 1.0.0
+ */
+module.exports = grunt => {
+	return {
+		options         : {
+			context: {
+				version: grunt.file.readJSON('package.json').version
+			}
+		},
+		versionInitiator: {
+			src : 'docs/scripts/initiators/version.initiator.js',
+			dest: '.tmp/release/version.initiator.js'
+		}
+	};
+};